pyproject.toml
[project]
name = "PyDocuSaurus"
version = "0.0.1"
description = "Generates Docusaurus Style Markdown documentation from Python module dosctrings"
authors = [
    {name = "Asthestarsfalll"},
]
dependencies = ["docstring-parser-fork>=0.0.12", "astor>=0.8.1"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "CC-0"}

[project.scripts]
pdocs = "PyDocuSaurus.generate:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.urls]
documentation = "https://github.com/criccomini/python-docstring-markdown/blob/main/DOCUMENTATION.md"
repository = "https://github.com/criccomini/python-docstring-markdown"
issues = "https://github.com/criccomini/python-docstring-markdown/issues"

[tool.pdm]
distribution = true

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "black>=25.1.0",
    "isort>=6.0.0",
    "pylint>=3.3.4",
    "pyright>=1.1.394",
    "autoflake>=2.3.1",
]
